Migrate to Gradle build (with Android Studio)
This commit migrates all files in the 'ChordReader' directory to the
'app' directory and switches to using Gradle as the build system. This
was done by importing the ChordReader directory into Android Studio
(ver. 3.2.1), then copying the resulting files into this tree. There
wa a bit of git-fu in there to get Git to recognize this as a simple
rename for most of the files.

This commit also removes the ChordReaderTest directory. It looks like
it hadn't been committed to in quite some time, and any function it
served should be covered by 'build variants' in Gradle.