Increase SDK version to 11 to get App bar to show
When downloading this app from f-droid, I realized that I couldn't get
to the menu. This makes the app pretty much unusable. By upgrading to
SDK 11, Android will default to adding a title bar (the 'App bar'),
which includes a menu button (I believe this is called the 'overflow
menu').

This was the most expedient way to get the app to function on
my device, and in no way indicates a desire to not support devices on
SDKs lower than 11. If a satisfactory way to support old devices,
while also displaying the App bar on newer devices is discovered, I'll
be happy to lower the target SDK back down.