Remove progress dialog on Open File
While browsing through the code I realized that there was an
unnecessary 1s delay while opening a file. On my device, and I assume
most modern devices, the chord text loads so fast, that no progress
dialog is needed. Removing it entirely makes the experience much more
seemless.

Perhaps this could be altered to check for device level and show the
progress bar only for 'older' devices. Not sure how to benchmark
this. But basically any device that takes longer than 1s to load the
text could show the progress bar. In which case the forced 1s delay
could be removed.