Add "New File" menu option
Sometimes I want to record a song that I know from memory or from a
physical text source. Before this commit, the only way to add a new
song is to search the web for a song and save it as a file.

This commit adds a menu item labeled 'New File', that allows the user
to create a new empty file to record a new song in. After prompting
for the filename, it places the user in view mode on the new file and
opens an edit dialog for it.