Import AlertDialog instead of AlertDialog.Builder
When reading through this file, I was initially very confused by the
use of the "Builder" class. I didn't know what it was building. After
doing some reading about the Builder design pattern, it seems like the
common usage of this pattern is to import just the parent class and
explicitly reference ParentClass.Builder when using it. This preserves
the context of which class you're instantiating with the Builder. It
also allows you to use builders for multiple classes in the same
file.

This commit switches to importing AlertDialog instead of
AlertDialog.Builder, and switches all references to Builder to
AlertDialog.Builder in the code. This make the code much more readable
in my opinion.