Remove unused field `regexString` and its accessors
According to Android Studio, the getter is not being accessed
from anywhere. When removing it, Android Studio indicates that
the field can be converted to a local variable inside the
setter, since is not being used anywhere else.

This makes it apparent that the whole field is not being used
(only being set at some point, never read), so it can be removed.