Allow for transpose settings to be changed by tapping text views
This change allows users to not only use the `SeekBar` when
changing the transpose settings, but also to tap on the "min"
and "max" `TextView`s.

From my experience, it can be difficult to select the correct
"progress" for the `SeekBar` when dragging and dropping. Tapping
the `TextView`s to in- or decrease the values by one should
be much easier.